week 3
Akash went to the market to buy N bulbs. He found two shops, shop A and B, where
bulbs were being sold in lots. He can buy any number of the complete lot(s) but not loose
bulbs. He is confused with the price and wants you to figure out the minimum cost to buy
exactly N bulbs.
Input Format:
 The first line of the input consists of an integer – N, representing the total number of
bulbs that Akash wants to buy.
 The second line consists of two space-separated positive integers – M1 and P1,
representing the number of bulbs in a lot and the lot’s price at shop A, respectively.
 The third line consists of two space-separated positive integers-M2 and P2,
representing the number of bulbs in a lot’s price at shop B, respectively.
Output Format:
 Print a positive integer representing the minimum price at which Akash can buy the